moved API fixture into test_setup
