+++ app/spec/javascript/test_setup.js
@@ -36,6 +36,16 @@ vi.mock('load-script', () => {
   }
 })
 
+const MOCK_PINWHEEL_AUTH_OBJECT = { token: 'test-token' };
+vi.mock('@js/utilities/api', async () => {
+  const apiModule = await vi.importActual('@js/utilities/api')
+  return {
+      ...apiModule,
+      trackUserAction: vi.fn((eventName, eventPayload) => Promise.resolve()),
+      fetchToken: vi.fn(() => Promise.resolve(MOCK_PINWHEEL_AUTH_OBJECT)),
+  }
+})
+
 // Reset all mocks before each test
 beforeEach(() => {
   // Set up Stimulus
